Move drawing profile to profile module
